% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CANS_DBSCAN.R
\name{CANS_DBSCAN}
\alias{CANS_DBSCAN}
\title{Cluster Analysis of Nucleotide Sequences. Density-based Spatial Clustering of
Applications with Noise (DBSCAN)}
\usage{
CANS_DBSCAN(sequences, e, k_min = 3, method_stringdistmatrix = "cosine")
}
\arguments{
\item{sequences}{List of nucleotide sequences; have to be in format \code{matrix}.}

\item{e}{e-neighborhood size.}

\item{k_min}{The minimum number of neighbors required for an object
to be considered primary. Typically selected from a range of 3 to 9. The default is 3.}

\item{method_stringdistmatrix}{Method for distance calculation. The default is "cosine", see \code{\link[=stringdist]{stringdist()}} - methods.}
}
\value{
A list of objects and cluster numbers to which they relate,
the \verb{number of clusters} and the \code{distance} matrix.
}
\description{
The function performs cluster analysis of nucleotide sequences
using the DBSACN method.It takes as input a list of nucleotide sequences \code{sequences},
e-neighborhood size \code{e}, the minimum number of neighbors required for an object
to be considered primary \code{k_min} and method for distance calculation \code{method_stringdistmatrix}.
}
\examples{
result <- CANS_DBSCAN(CANS_imitation(number_of_clusters = 3), e = 4, k_min = 5, "cosine")
result <- CANS_DBSCAN(CANS_imitation(number_of_clusters = 4), e = 4, k_min = 3, "qgram")
}
\seealso{
\code{\link[=stringdist]{stringdist()}}.
}
