% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CANS_hierarchical_clustering.R
\name{CANS_hierarchical_clustering}
\alias{CANS_hierarchical_clustering}
\title{Cluster Analysis of Nucleotide Sequences. Hierarchical clustering method.}
\usage{
CANS_hierarchical_clustering(
  sequences,
  number_of_clusters,
  method_stringdistmatrix = "cosine",
  method_hclust = "complete"
)
}
\arguments{
\item{sequences}{List of nucleotide sequences; have to be in format \code{matrix}.}

\item{number_of_clusters}{The number of clusters.}

\item{method_stringdistmatrix}{Method for distance calculation. The default is "cosine", see \code{\link[=stringdist]{stringdist()}} - methods.}

\item{method_hclust}{The agglomeration method to be used. The default is "complete", see \code{\link[=hclust]{hclust()}} - metrics.}
}
\value{
A list of objects and cluster numbers to which they relate, cophenetic coefficient
}
\description{
The function performs cluster analysis of nucleotide sequences
using the hierarchical clustering method. It takes as input a list of nucleotide
sequences \code{sequences}, the number of clusters \code{number_of_clusters}, method for
distance calculation \code{method_stringdistmatrix}and cluster linking method \code{method_hclust}.
}
\examples{
\dontrun{
result <- CANS_hierarchical_clustering(CANS_imitation(number_of_clusters = 4),
 number_of_clusters = 4, "cosine")
result <- CANS_hierarchical_clustering(CANS_imitation(number_of_clusters = 3),
 number_of_clusters = 3, "jw")
 }
}
\seealso{
\code{\link[=stringdistmatrix]{stringdistmatrix()}}, \code{\link[=hclust]{hclust()}}, \code{\link[=dendextend]{dendextend()}}, \code{\link[=stats]{stats()}}.
}
