% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CANS_imitation.R
\name{CANS_imitation}
\alias{CANS_imitation}
\title{Cluster Analysis of Nucleotide Sequences. Imitation model.}
\usage{
CANS_imitation(
  number_of_sequences = 100,
  sequences_length = 80,
  number_of_clusters = 1,
  nucleotides = c("A", "T", "C", "G"),
  randomness = 0.4
)
}
\arguments{
\item{number_of_sequences}{The number of sequences in one cluster. The default is 100.}

\item{sequences_length}{The length of each sequence. The default is 100.}

\item{number_of_clusters}{The number of clusters. The default is 1.}

\item{nucleotides}{The alphabet of characters. The default is c("\code{A}", "\code{T}", "\code{C}", "\code{G}").}

\item{randomness}{Determines the maximum proportion of differences between the
elements of each cluster. Varies from 0 to 1. The default is 0.4.}
}
\value{
A list of nucleaotide sequences.
}
\description{
Generates a list of clusters of nucleotide sequences  for performing
cluster analysis. Takes parameters as input: \verb{number of sequences}, their \code{length},
\verb{number of clusters}, alphabet of characters \code{nucleotides} and \code{randomness} parameter.
The function creates random starting nucleotide sequences for each cluster,
as well as a vector of random nucleotide positions in the given sequences. Generates
new sequences based on previously generated ones by changing nucleotides at
certain positions. Thus, clusters of similar sequences are formed, with a share
of similar symbols of at least 1 - \code{randomness}.
}
\examples{
\dontrun{
sequences <- CANS_imitation(number_of_sequences = 100, sequences_length = 80,
number_of_clusters = 3, randomness = 0.3)

sequences <- CANS_imitation()
}
}
