% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CANS_k_means.R
\name{CANS_k_means}
\alias{CANS_k_means}
\title{Cluster Analysis of Nucleotide Sequences. K-Means method.}
\usage{
CANS_k_means(
  sequences,
  k,
  e = 0.001,
  m_max = 100,
  method_stringdistmatrix = "cosine"
)
}
\arguments{
\item{sequences}{List of nucleotide sequences; have to be in format \code{matrix}.}

\item{k}{The number of clusters.}

\item{e}{Maximum value of the difference in clustering quality between two sequential iterations.}

\item{m_max}{The maximum number of iterations allowed.}

\item{method_stringdistmatrix}{Method for distance calculation. The default is "cosine", see \code{\link[=stringdist]{stringdist()}} - methods.}
}
\value{
A list of objects and cluster numbers to which they relate,
indexes of central objects, as well as the number of iterations performed.
}
\description{
The function performs cluster analysis of nucleotide sequences
using the k-means method. It takes as input a list of nucleotide sequences \code{sequences},
the number of estimated clusters \code{k}, clustering accuracy \code{e} and the maximum
allowed number of iterations \code{m_max} and method for distance calculation \code{method_stringdistmatrix}.
}
\examples{
result <- CANS_k_means(CANS_imitation(number_of_clusters = 2), k = 2, "qgram")
result <- CANS_k_means(CANS_imitation(number_of_clusters = 4), k = 4, e = 0.00001, "jw")
}
\seealso{
\code{\link[=stringdistmatrix]{stringdistmatrix()}}, \code{\link[=kmeans]{kmeans()}}.
}
