% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CANS_spectral.R
\name{CANS_spectral}
\alias{CANS_spectral}
\title{Cluster Analysis of Nucleotide Sequences. Spectral clustering.}
\usage{
CANS_spectral(
  sequences,
  number_of_neighbors = 11,
  method_stringdistmatrix = "cosine",
  number_of_clusters = 2
)
}
\arguments{
\item{sequences}{List of nucleotide sequences; have to be in format \code{matrix}.}

\item{number_of_neighbors}{The number of significant neighboring objects.
Recommended to select at least 10. The default is 11.}

\item{method_stringdistmatrix}{Method for distance calculation. The default is "cosine", see \code{\link[=stringdist]{stringdist()}} - methods.}

\item{number_of_clusters}{The number of clusters.}
}
\value{
A list of objects and cluster numbers to which they relate,
eigenvectors on which the transition to the reduced feature space is performed.
}
\description{
The function performs cluster analysis of nucleotide sequences
using the spectral method. It takes as input a list of nucleotide sequences \code{sequences},
the number of significant neighboring objects \code{number_of_neighbors}, method for
distance calculation \code{method_stringdistmatrix} and the number of estimated clusters
\code{number_of_clusters}.
}
\examples{
result <- CANS_spectral(CANS_imitation(number_of_clusters = 2),
    method_stringdistmatrix = "qgram", 2)
result <- CANS_spectral(CANS_imitation(number_of_clusters = 4),
    method_stringdistmatrix = "cosine", 4)
}
\seealso{
\code{\link[=stringdist]{stringdist()}}, \code{\link[=eigen]{eigen()}}, \code{\link[=kmeans]{kmeans()}}.
}
